Split GraphQL schema definition into separate modules
