+++ packages/launcher/src/renderer/launcher/Launcher.js
@@ -87,7 +87,7 @@ class Launcher extends Component<Props, State> {
 
 const LauncherRelayContaier = createFragmentContainer(Launcher, {
   identities: graphql`
-    fragment Launcher_identities on IdentitiesQuery {
+    fragment Launcher_identities on Identities {
       ownUsers {
         defaultEthAddress
         localID
